frontend/fastapi_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv 